added CRUD operations for exams

+++ src/main/java/com/java/jkpot/dao/CountersDAO.java
@@ -2,5 +2,5 @@ package com.java.jkpot.dao;
 
 public interface CountersDAO {
 
-	long getNextSequenceOfField(String fieldName);
+	int getNextSequenceOfField(String fieldName);
 }
